Working on the cucumber tests
Learning how to fire up the database correctly is tricky.

I need to build a class that has the methods for talking to my database
like I want. Probably as an actor, so that it can keep the in memory
database open for testing.

Or use a file backed database like with other projects.
Or just do it against postgresql the whole time, but that prohibits
simple testing. Should do it with H2, and make it work.